ios-bootcamp-challenge: implement generics on PokeAPI

+++ iOS Bootcamp Challenge/Controllers/DetailViewController.swift
@@ -163,7 +163,7 @@ class DetailViewController: UIViewController {
             label.backgroundColor = .white.withAlphaComponent(0.30)
             label.layer.cornerRadius = 7.0
             label.layer.masksToBounds = true
-            let paddedWidth = label.intrinsicContentSize.width + padding
+            let paddedWidth = label.intrinsicContentSize.width + CGFloat(padding)
             label.widthAnchor.constraint(equalToConstant: paddedWidth).isActive = true
             typesStackView.addArrangedSubview(label)
         }
